// dashboard-next/apps/web/src/features/tuf/scripts/types.ts
export type StepStatus = 'ready' | 'in-progress' | 'success' | 'error' | 'waiting' | 'disabled'

export interface TufHistoryEntry {
  id: string
  timestamp: string
  appName: string
  operation: 'generate' | 'bootstrap' | 'publish' | 'unsign' | 'update-config' | 'root-meta-update' | 'metadata-update'
  status: 'success' | 'failed' | 'pending'
  taskId?: string
  version?: string
}

export type TaskState = 'PENDING' | 'SUCCESS' | 'FAILURE'

export interface TaskData {
  task_id: string
  state: TaskState
  result?: {
    message: string
    status?: boolean
    task?: string
    last_update?: string
    error?: string
  }
}

export interface TufConfig {
  bootstrap: string
  role_expiration: number
  root_expiration: number
  root_num_keys: number
  root_threshold: number
  snapshot_expiration: number
  snapshot_num_keys: number
  snapshot_threshold: number
  targets_expiration: number
  targets_num_keys: number
  targets_online_key: number
  targets_threshold: number
  timestamp_expiration: number
  timestamp_num_keys: number
  timestamp_threshold: number
}